add small fixes + add cov html report
