+++ .gitignore
@@ -37,7 +37,6 @@ pip-log.txt
 pip-delete-this-directory.txt
 
 # Unit test / coverage reports
-htmlcov/
 .tox/
 .nox/
 .coverage
@@ -158,3 +157,16 @@ cython_debug/
 #  and can be added to the global gitignore or merged into this file.  For a more nuclear
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 .idea/
+/htmlcov/class_index.html
+/htmlcov/coverage_html_cb_6fb7b396.js
+/htmlcov/favicon_32_cb_58284776.png
+/htmlcov/function_index.html
+/htmlcov/keybd_closed_cb_ce680311.png
+/htmlcov/status.json
+/htmlcov/style_cb_8e611ae1.css
+/htmlcov/z_f8996bc1f023edad___init___py.html
+/htmlcov/z_f8996bc1f023edad_albums_py.html
+/htmlcov/z_f8996bc1f023edad_bottle_py.html
+/htmlcov/z_f8996bc1f023edad_main_py.html
+/htmlcov/z_f8996bc1f023edad_number_py.html
+/htmlcov/z_f8996bc1f023edad_students_py.html
