+++ pyproject.toml
@@ -17,3 +17,12 @@ dependencies = [
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
 build-backend = "poetry.core.masonry.api"
+
+[tool.pytest.ini_options]
+pythonpath = ["."]
+testpaths = ["lesson1/tests"]
+
+[tool.coverage.run]
+source = ["lesson1"]
+omit = ["*/tests/*"]
+
